fix(dagent): add import container logs parsing & send
